Backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
coverage==7.6.10
dill==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
